complete bill model and migrate to db

+++ migrations/20180730005408-create-user.js
@@ -26,7 +26,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       monthlySalary: {
-        type: Sequelize.MONEY,
+        type: Sequelize.INTEGER,
         allowNull: false
       },
       createdAt: {
